solving problems

+++ miniRt/new-minirt/src/utilities/rgbcolor.h
@@ -6,7 +6,7 @@
 /*   By: irifarac <irifarac@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/10 11:01:51 by irifarac          #+#    #+#             */
-/*   Updated: 2023/03/10 11:04:51 by irifarac         ###   ########.fr       */
+/*   Updated: 2023/03/20 10:03:55 by irifarac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 
 typedef struct s_rgbcolor
 {
-	float	r;
-	float	g;
-	float	b;
+	int	r;
+	int	g;
+	int	b;
 }	t_rgb;
 
 #endif
